+++ test/elven_gard/network/endpoint/ranch_test.exs
@@ -46,6 +46,20 @@ defmodule ElvenGard.Network.Endpoint.RanchTest do
     transport_options: [certfile: @tls_certfile, keyfile: @tls_keyfile]
   )
 
+  Application.put_env(:elvengard_network, __MODULE__.ShutdownEndpoint,
+    adapter: ElvenGard.Network.Endpoint.Adapters.Ranch,
+    ip: {127, 0, 0, 1},
+    listener_name: :ranch_shutdown_endpoint,
+    port: 0,
+    socket_handler: __MODULE__.ShutdownSocketHandler,
+    transport: :tcp
+  )
+
+  Application.put_env(:elvengard_network, __MODULE__.ShutdownSocketHandler,
+    network_codec: __MODULE__.HaltCodec,
+    packet_handler: __MODULE__.HaltHandler
+  )
+
   defmodule HaltPacket do
     defstruct [:data]
   end
@@ -153,6 +167,10 @@ defmodule ElvenGard.Network.Endpoint.RanchTest do
     use ElvenGard.Network.Endpoint, otp_app: :elvengard_network
   end
 
+  defmodule ShutdownEndpoint do
+    use ElvenGard.Network.Endpoint, otp_app: :elvengard_network
+  end
+
   defmodule InitStopSocketHandler do
     use ElvenGard.Network.SocketHandler
 
@@ -203,6 +221,24 @@ defmodule ElvenGard.Network.Endpoint.RanchTest do
     def handle_halt(reason, %{assigns: %{link: _}} = socket) do
       %{assigns: %{link: link}} = socket
       send(link, {:handle_halt, reason})
+      {:ok, socket}
+    end
+  end
+
+  defmodule ShutdownSocketHandler do
+    use ElvenGard.Network.SocketHandler
+
+    ## SocketHandler callbacks
+
+    @impl true
+    def handle_init(socket) do
+      send(Process.whereis(__MODULE__.Observer), :shutdown_init)
+      {:ok, socket}
+    end
+
+    @impl true
+    def handle_halt(reason, socket) do
+      send(Process.whereis(__MODULE__.Observer), {:shutdown_halt, reason})
       {:ok, socket}
     end
   end
@@ -269,6 +305,28 @@ defmodule ElvenGard.Network.Endpoint.RanchTest do
 
       :ok = :ssl.close(socket)
       assert_receive {:handle_halt, :closed}
+    end
+  end
+
+  describe "listener shutdown" do
+    test "closes active connections and runs halt cleanup" do
+      Process.register(self(), ShutdownSocketHandler.Observer)
+      child_spec = ShutdownEndpoint.child_spec([])
+      start_supervised!(child_spec)
+
+      {:ok, socket} =
+        :gen_tcp.connect(
+          {127, 0, 0, 1},
+          ShutdownEndpoint.get_port(),
+          [:binary, active: false]
+        )
+
+      assert_receive :shutdown_init
+      :ok = stop_supervised(child_spec.id)
+
+      assert_receive {:shutdown_halt, :closed}
+      assert {:error, :closed} = :gen_tcp.recv(socket, 0)
+      refute_received {:shutdown_halt, :closed}
     end
   end